Add NavBar & NavMenu
In this commit we create a bare-bones navbar for the app, and a nav menu (appearing when a hamburger menu is clicked). The nav menu has a single select field, controlled by the nav menu's state. To make this field (and others) fully functional, we'll need to lift this state up and re-render the markers according to user selections).

For this update we add a couple of new dependencies:

* Material UI for our components and styling
* Lodash for a few utility functions

We also update api_is_loaded to avoid updating the bounds of the map unless there is map data available.

Material UI's CssBaseline component proves really useful in helping us to normalize our CSS.