Filter Meetings By Yearly Meeting
This commit updates our yearly_meetings dropdown to allow users to filter meetings on the map by yearly meeting.

It should be pretty easy from this point to do this kind of filtering with any other type of value that is stored as a string on meeting.

Because of the way this is implemented, we'll need to do a little more thinking to make this work for our number- and boolean-typed values (such as lgbt-affirming, or latitude/logitude).

As a part of this update, we renamed apiIsLoaded to updateMapBounds, as this more accurately reflects what this function is doing. updateMapBounds is now called in two places - on initial map load, and then again every time appState.filteredMeetings is updated.