Add All String Fields As Dropdowns in Main NavMenu
This commit has us ignoring a pretty gnarly typescript issue that isn't allowing us to use strings as an index type for Meeting. Not quite sure how to resolve this, so ignoring for now; everything  seems to work fine otherwise.

One issue: At present if you reduce filteredMeetings to an empty array through your selections, the map will not update (per line 31 of update_map_bounds). This isn't a great user experience, as the map doesn't reflect the state of the dropdowns.

We need to think about what we want the behavior to be here. Either presenting an empty map when the options exclude all meetings, or to disallow selecting an option that would exclude all possible meetings.

Also, added 'country' as a field for meetings, since people will definitely want to be able to search by nation.