Add React Router to Render Multiple Views for Site
With React Router, we now have four additional "pages" in addition to the main map (index) view: About, Contact, FAQ, and Info

This change was a bit tricker than anticipated with regards to the MainMap and updating map state. We had a feedback loop between updateMapBounds and our useEffect hook in App that fires every time appState.filteredMeetings is updated.

In order to avoid this feedback loop, we reorganized our state. This may not be the best long-term solution, but it works for now and can be refactored to be more useable as we move forward.

Several things obviously need doing at this point:

* The various "pages" need to be styled and have images and forms added
* The pages need to be accessible via some kind of links or dropdowns from the nav bar
* The NavMenu should only be available in the MainMap view
* Some of our dropdowns - in particular worship_style and accessibility - don't seem to be working properly right now