Fix Meetings Display & Move Filter Logic from App
Our meetings weren't showing up properly on the main map because we were displaying from the meetings rather than the filtered meetings.

Our logic for filtering meetings doesn't belong in the view layer, so we abstracted it out into a util function. This will be even more important as we continue to refine and expand on our filter logic.