Load Meetings Data from JSON File Rather Than API
After all that work rigging up our backend, it really doesn't seem like Firebase is the best solution for the problem we're trying to solve. Namely, querying ~1200 records (with a daily 50k cap for querying records). It just doesn't scale.

Fortunately, simply keeping our meetings data in a JSON file and loading it directly as a front-end asset seems to work pretty well. And it's a lot faster.

Other updates in this commit:

* Greatly improve performance by setting our opening map to load a random yearly meeting, rather than every meeting in the collection. (The user can still choose to view all meetings, if they wish.)
* Get rid of the About static page and have its URL redirect to the FAQ page. (FAQ is stronger and About was redundant.)
* Get rid of accessibility and zip code dropdowns. They're just not that useful; perhaps we'll reimplement zip codes later as a typeable search field
* Redo styled components in Material UI style of CSS management
* Sort titles alphabetically so they make sense in the dropdowns
* Smooth out some edge cases with getFilterMeetings