Don't Update Filter Select If Meetings Not Updated
Previously you could change a Filter Meetings dropdown in the NavMenu, and it would update the dropdown selection, regardless of whether it was a valid option that updated the meetings state. For example, you could select "LGBT-Affirming: false" for meetings that were already filtered by a yearly meeting where every local meeting was LGBT-affirming.

With this update, you can't select options that won't update the state. So if for that LGBT-affirming yearly meeting, you seek to filter by LGBT-Affirming: false, the meetings won't update, and neither will the select option.

Also: Change records for meetings in "Pacific YM" to be instead from "Pacific Yearly Meeting" to avoid name collisions with "North Pacific YM"