Center Markers On Geolocation At All Zoom Levels
Meeting markers must be repositioned manually because google-map-react positions the marker according to its upper-left-hand corner, rather than its center. Without this styling, the markers change positions depending on the level of zoom.

For more information, see this bug report: https://github.com/google-map-react/google-map-react/issues/796

This might be a bug that's worth fixing and submitting a PR to the library