Fix Bug with Meetings Icon Popover Shadow Position
We had this really strange issue where the circular box shadow that Material-UI includes on hover for IconButton was positioned according to a different rule from the HomeIcon, which was repositioned to not move when zoomed (see previous commit).

What made this tricky was that it was not clear how we could simply style the IconButton to match positions with its child HomeIcon. Adding the same styling to both parent and child elements ended up replicating the zoom issue we were seeing before.

What finally solved this issue was applyig the marker styles to the IconButton and - rather than nesting HomeIcon inside IconButton - passing HomeIcon as a children prop to IconButton.

Very strange thing, but seems to be working great now. Still not sure why we couldn't style the parent element and nest the child element rather than passing it as a prop.

In addition to this change, also update our meetings data to include Sierra Cascades YM and fix a few errata.