Move FIrebase Assets Into Their Own Directory
For the time being, we're not going to be using firebase for this project. However, if we get into users and authentication, Firebase could once again be quite useful. We're going to leave this code here for the time being, for reference, but keep it clearly segregated from the client code, which is our working app at this point.