Give Option to Search By Single Criterion When Attempted Search Invalid
It's common that a user will enter a combination of criteria, only to find the last one does not return any results. Previously, in a case like this, the form would just snap back to its previous state and give no explanation. Even worse, the only way to get out of this state and perform the desired search, was to click out of all criteria, which forced the user to return all results - an intensive operation that takes several seconds.

The new behavior is, when the user makes a selection that would return no meetings, they see a flash alert in the bottom left hand corner of the viewport, letting them know that their last selection was invalid. At the same time, a red button appears in the NavMenu which, if clicked, will trigger a new search on _only_ that single criterion. This allows the user to search, for example, for all meetings in a particular state, without needing to un-select all other criteria first.