NavMenu Displaces MainMap/NavBar Rather Than Covering Them
In order to make this work, we needed to move NavMenu up to the same level as NavBar. These two now live together in SiteNav, and share state with MainMap (which, along with NavBar, gets out of NavMenu when it opens).

A bit of a problem here: Each time we open or close the menu, we have to updateMapBounds again.

Not clear what way we might find to work around this, but seems like we'd like to avoid this extra overhead if we could. Not to mention it doesn't look great.