Manage MainMap State in MainMap Component
We had raised up a lot of MainMap's business to App for no reason that is apparent at this point.

In addition to lowering MainMap's state back down, we also got rid of redundant logic in App that was meant to handle re-bounding the map, but which in reality seemed to serve no purpose.