NavMenu Covers Full Width In Mobile View
The NavMenu worked well on desktop view, but its width quickly became problematic at tablet and mobile.

This update adds new behavior for NavMenu at widths below 960px, causing the menu to "cover" the entire screen.

It's questionable whether using the window.onresize event is the best way to watch for changes in the viewport, but it's the best way we've found for the time being.