Fix Bug with Multi-Select Attribute Meeting Filtering
Our logic wasn't correct. It was also ugly. Hopefully the code is not only improved functionally, but at least slightly more readable now.

The key changes here are on lines 41-42. The rest is variable name updates to hopefully make this filtering logical more grokkable.

This is still one of the hardest parts of this codebase to read, so we should keep making improvements as we go.