Fix Update Meeting Form Bug on Safari
We were seeing weird behavior where form elements would change order whenever state was updated. (The input/select that was updated would move to the bottom.)

Turns out there were a couple problems here:

First, we didn't have a unique key for our elements, and this ended up with us shifting focus when the input moved. This was an easy fix.

Second, Object.keys() was returning inconsistently ordered elements. Consistent and ordered on Chrome and Firefox, but Safari would change the order of the elements when the object state was updated. Turns out, we can't rely on Object.keys() for order. For now, we'll just .sort it and be done with it. If we want a more specific order later, we may need to hard-code, but would like to avoid that.

Also: Got rid of an extra CircularProgress spinner in UpdateMeetings.