Migrate to @vis.gl/react-google-maps
We are migrating away from the no-longer-maintained GoogleMapReact to the blessed package maintained by the good people at Vis.gl.

For ease of migration, we are using the <Marker /> component, but we probably want to migrate to <AdvancedMarker /> soon, as this is the standard and <Marker /> will likely be deprecated in the future.